Move Event and Command Model into module
Move the EventModel to Model package into event module #172
Move CommandModel to model package into command module #173
