+++ tests/test_common/test_broker.py
@@ -12,14 +12,10 @@ from typing import (
 
 from minos.common import (
     Aggregate,
-    Command,
-    Event,
     MinosBaseBroker,
-    MultiTypeMinosModelSequenceException,
 )
 from tests.aggregate_classes import (
     Car,
-    Owner,
 )
 
 
@@ -40,24 +36,6 @@ class TestMinosBaseBroker(unittest.IsolatedAsyncioTestCase):
     async def test_send_one(self):
         broker = MinosBroker("CarAdded")
         self.assertEqual(None, await broker.send_one(Car(1, 1, 3, "red")))
-
-
-class TestEvent(unittest.TestCase):
-    def test_avro_serialization(self):
-        event = Event("CarCreated", [Car(1, 1, 3, "blue"), Car(2, 1, 5, "red")])
-        decoded_event = Event.from_avro_bytes(event.avro_bytes)
-        self.assertEqual(event, decoded_event)
-
-    def test_items_raises(self):
-        with self.assertRaises(MultiTypeMinosModelSequenceException):
-            Event("CarCreated", [Car(1, 1, 3, "blue"), Owner(2, 1, "Foo", "Bar")])
-
-
-class TestCommand(unittest.TestCase):
-    def test_avro_serialization(self):
-        command = Command("CarCreated", [Car(1, 1, 3, "blue"), Car(2, 1, 5, "red")], "reply_fn()")
-        decoded_command = Command.from_avro_bytes(command.avro_bytes)
-        self.assertEqual(command, decoded_command)
 
 
 if __name__ == "__main__":
